requirements.txt
marvin==0.9.0
nltk==3.8.1
beautifulsoup4==4.12.2
prefect==2.10.9
readability-lxml==0.8.1
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
aiosqlite==0.19.0
alembic==1.11.0
anyio==3.6.2
appnope==0.1.3
apprise==1.4.0
argcomplete==3.0.8
asgi-lifespan==2.1.0
asttokens==2.2.1
async-timeout==4.0.2
asyncpg==0.27.0
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
black==23.3.0
bleach==6.0.0
cachetools==5.3.0
certifi==2023.5.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
colorama==0.4.6
coolname==2.2.0
courlan==0.9.2
croniter==1.3.14
cryptography==40.0.2
cssselect==1.2.0
dataclasses-json==0.5.7
datamodel-code-generator==0.19.0
dateparser==1.1.8
decorator==5.1.1
diskcache==5.6.1
dnspython==2.3.0
docker==6.1.2
duckduckgo-search==2.9.5
email-validator==2.0.0.post2
executing==1.2.0
fake-useragent==1.1.3
fastapi==0.95.2
frozenlist==1.3.3
fsspec==2023.5.0
genson==1.2.2
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.18.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.59.0
greenlet==2.0.2
griffe==0.27.5
h11==0.14.0
h2==4.1.0
hpack==4.0.0
htmldate==1.4.3
httpcore==0.17.0
httplib2==0.22.0
httpx==0.24.0
hyperframe==6.0.1
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
inflect==5.6.2
ipython==8.13.2
isort==5.12.0
jedi==0.18.2
jellyfish==0.11.2
Jinja2==3.1.2
joblib==1.2.0
jsonpatch==1.32
jsonpointer==2.3
jsonschema==4.17.3
jsonschema-spec==0.1.4
jusText==3.0.0
kubernetes==26.1.0
langchain==0.0.153
langcodes==3.3.0
lazy-object-proxy==1.9.0
linkify-it-py==2.0.2
lxml==4.9.2
Mako==1.2.4
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
marvin==0.9.0
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.5
mdurl==0.1.2
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.6
networkx==3.1
nltk==3.8.1
numexpr==2.8.4
numpy==1.24.3
oauthlib==3.2.2
openai==0.27.6
openapi-schema-pydantic==1.2.4
openapi-schema-validator==0.3.4
openapi-spec-validator==0.5.1
orjson==3.8.12
packaging==23.1
parso==0.8.3
pathable==0.4.3
pathspec==0.11.1
pendulum==2.1.2
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.5.1
prance==0.22.2.22.0
prefect==2.10.9
prompt-toolkit==3.0.38
protobuf==4.23.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.7
Pygments==2.15.1
pyparsing==3.0.9
pyperclip==1.8.2
pyrsistent==0.19.3
PySnooper==1.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
python-slugify==8.0.1
pytz==2023.3
pytzdata==2020.1
PyYAML==6.0
readability-lxml==0.8.1
readchar==4.0.5
regex==2023.5.5
requests==2.30.0
requests-oauthlib==1.3.1
rich==13.3.5
rsa==4.9
ruamel.yaml==0.17.26
ruamel.yaml.clib==0.2.7
segtok==1.5.11
simpleeval==0.9.13
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a34
sqlitedict==2.1.0
sqlmodel==0.0.8
stack-data==0.6.2
starlette==0.27.0
tabulate==0.9.0
tenacity==8.2.2
text-unidecode==1.3
textual==0.24.1
tiktoken==0.4.0
tld==0.13
toml==0.10.2
tomli==2.0.1
tqdm==4.65.0
trafilatura==1.6.0
traitlets==5.9.0
typer==0.9.0
typing-inspect==0.8.0
typing_extensions==4.5.0
tzlocal==5.0.1
uc-micro-py==1.0.2
ulid-py==1.1.0
uritemplate==4.1.1
urllib3==1.26.15
uvicorn==0.22.0
wcwidth==0.2.6
webencodings==0.5.1
websocket-client==1.5.1
websockets==11.0.3
wikipedia==1.4.0
xxhash==3.2.0
yake==0.4.8
yarl==1.9.2
zipp==3.15.0
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
aiosqlite==0.19.0
alembic==1.11.0
anyio==3.6.2
appnope==0.1.3
apprise==1.4.0
argcomplete==3.0.8
asgi-lifespan==2.1.0
asttokens==2.2.1
async-timeout==4.0.2
asyncpg==0.27.0
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
black==23.3.0
bleach==6.0.0
cachetools==5.3.0
certifi==2023.5.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
colorama==0.4.6
coolname==2.2.0
courlan==0.9.2
croniter==1.3.14
cryptography==40.0.2
cssselect==1.2.0
dataclasses-json==0.5.7
datamodel-code-generator==0.19.0
dateparser==1.1.8
decorator==5.1.1
diskcache==5.6.1
dnspython==2.3.0
docker==6.1.2
duckduckgo-search==2.9.5
email-validator==2.0.0.post2
executing==1.2.0
fake-useragent==1.1.3
fastapi==0.95.2
frozenlist==1.3.3
fsspec==2023.5.0
genson==1.2.2
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.18.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.59.0
greenlet==2.0.2
griffe==0.27.5
h11==0.14.0
h2==4.1.0
hpack==4.0.0
htmldate==1.4.3
httpcore==0.17.0
httplib2==0.22.0
httpx==0.24.0
hyperframe==6.0.1
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
inflect==5.6.2
ipython==8.13.2
isort==5.12.0
jedi==0.18.2
jellyfish==0.11.2
Jinja2==3.1.2
joblib==1.2.0
jsonpatch==1.32
jsonpointer==2.3
jsonschema==4.17.3
jsonschema-spec==0.1.4
jusText==3.0.0
kubernetes==26.1.0
langchain==0.0.153
langcodes==3.3.0
lazy-object-proxy==1.9.0
linkify-it-py==2.0.2
lxml==4.9.2
Mako==1.2.4
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
marvin==0.9.0
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.5
mdurl==0.1.2
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.6
networkx==3.1
nltk==3.8.1
numexpr==2.8.4
numpy==1.24.3
oauthlib==3.2.2
openai==0.27.6
openapi-schema-pydantic==1.2.4
openapi-schema-validator==0.3.4
openapi-spec-validator==0.5.1
orjson==3.8.12
packaging==23.1
parso==0.8.3
pathable==0.4.3
pathspec==0.11.1
pendulum==2.1.2
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.5.1
prance==0.22.2.22.0
prefect==2.10.9
prompt-toolkit==3.0.38
protobuf==4.23.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.7
Pygments==2.15.1
pyparsing==3.0.9
pyperclip==1.8.2
pyrsistent==0.19.3
PySnooper==1.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
python-slugify==8.0.1
pytz==2023.3
pytzdata==2020.1
PyYAML==6.0
readability-lxml==0.8.1
readchar==4.0.5
regex==2023.5.5
requests==2.30.0
requests-oauthlib==1.3.1
rich==13.3.5
rsa==4.9
ruamel.yaml==0.17.26
ruamel.yaml.clib==0.2.7
segtok==1.5.11
simpleeval==0.9.13
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a34
sqlitedict==2.1.0
sqlmodel==0.0.8
stack-data==0.6.2
starlette==0.27.0
tabulate==0.9.0
tenacity==8.2.2
text-unidecode==1.3
textual==0.24.1
tiktoken==0.4.0
tld==0.13
toml==0.10.2
tomli==2.0.1
tqdm==4.65.0
trafilatura==1.6.0
traitlets==5.9.0
typer==0.9.0
typing-inspect==0.8.0
typing_extensions==4.5.0
tzlocal==5.0.1
uc-micro-py==1.0.2
ulid-py==1.1.0
uritemplate==4.1.1
urllib3==1.26.15
uvicorn==0.22.0
wcwidth==0.2.6
webencodings==0.5.1
websocket-client==1.5.1
websockets==11.0.3
wikipedia==1.4.0
xxhash==3.2.0
yake==0.4.8
yarl==1.9.2
zipp==3.15.0
